backend/backend/requirements.txt
fastapi
python-dotenv
requests
spacy
uvicorn